Added prefix into config, added config into services for DI
